refactoring; add incoming banners to channel_list screen
