+++ app/screens/channel/index.tsx
@@ -3,15 +3,9 @@
 
 import {withDatabase} from '@nozbe/watermelondb/DatabaseProvider';
 import withObservables from '@nozbe/with-observables';
-import {combineLatest, distinctUntilChanged, of as of$, switchMap} from 'rxjs';
+import {of as of$, switchMap} from 'rxjs';
 
-import {observeIsCallsEnabledInChannel} from '@calls/observers';
-import {
-    observeCallsState,
-    observeChannelsWithCalls,
-    observeCurrentCall,
-    observeIncomingCalls,
-} from '@calls/state';
+import {observeCallStateInChannel, observeIsCallsEnabledInChannel} from '@calls/observers';
 import {Preferences} from '@constants';
 import {withServerUrl} from '@context/server';
 import {observeCurrentChannel} from '@queries/servers/channel';
@@ -29,37 +23,6 @@ type EnhanceProps = WithDatabaseArgs & {
 
 const enhanced = withObservables([], ({database, serverUrl}: EnhanceProps) => {
     const channelId = observeCurrentChannelId(database);
-
-    const isCallInCurrentChannel = combineLatest([channelId, observeChannelsWithCalls(serverUrl)]).pipe(
-        switchMap(([id, calls]) => of$(Boolean(calls[id]))),
-        distinctUntilChanged(),
-    );
-    const currentCall = observeCurrentCall();
-    const ccChannelId = currentCall.pipe(
-        switchMap((call) => of$(call?.channelId)),
-        distinctUntilChanged(),
-    );
-    const isInACall = currentCall.pipe(
-        switchMap((call) => of$(Boolean(call?.connected))),
-        distinctUntilChanged(),
-    );
-    const dismissed = combineLatest([channelId, observeCallsState(serverUrl)]).pipe(
-        switchMap(([id, state]) => of$(Boolean(state.calls[id]?.dismissed[state.myUserId]))),
-        distinctUntilChanged(),
-    );
-    const isInCurrentChannelCall = combineLatest([channelId, ccChannelId]).pipe(
-        switchMap(([id, ccId]) => of$(id === ccId)),
-        distinctUntilChanged(),
-    );
-    const showJoinCallBanner = combineLatest([isCallInCurrentChannel, dismissed, isInCurrentChannelCall]).pipe(
-        switchMap(([isCall, dism, inCurrCall]) => of$(Boolean(isCall && !dism && !inCurrCall))),
-        distinctUntilChanged(),
-    );
-    const showIncomingCalls = observeIncomingCalls().pipe(
-        switchMap((ics) => of$(ics.incomingCalls.length > 0)),
-        distinctUntilChanged(),
-    );
-
     const dismissedGMasDMNotice = queryPreferencesByCategoryAndName(database, Preferences.CATEGORIES.SYSTEM_NOTICE, Preferences.NOTICES.GM_AS_DM).observe();
     const channelType = observeCurrentChannel(database).pipe(switchMap((c) => of$(c?.type)));
     const currentUserId = observeCurrentUserId(database);
@@ -67,9 +30,7 @@ const enhanced = withObservables([], ({database, serverUrl}: EnhanceProps) => {
 
     return {
         channelId,
-        showJoinCallBanner,
-        isInACall,
-        showIncomingCalls,
+        ...observeCallStateInChannel(serverUrl, database, channelId),
         isCallsEnabledInChannel: observeIsCallsEnabledInChannel(database, serverUrl, channelId),
         dismissedGMasDMNotice,
         channelType,
