+++ app/screens/home/channel_list/channel_list.tsx
@@ -10,6 +10,7 @@ import Animated, {useAnimatedStyle, withTiming} from 'react-native-reanimated';
 import {type Edge, SafeAreaView, useSafeAreaInsets} from 'react-native-safe-area-context';
 
 import {refetchCurrentUser} from '@actions/remote/user';
+import FloatingCallContainer from '@calls/components/floating_call_container';
 import AnnouncementBanner from '@components/announcement_banner';
 import ConnectionBanner from '@components/connection_banner';
 import TeamSidebar from '@components/team_sidebar';
@@ -40,6 +41,7 @@ type ChannelProps = {
     coldStart?: boolean;
     currentUserId?: string;
     hasCurrentUser: boolean;
+    showIncomingCalls: boolean;
 };
 
 const edges: Edge[] = ['bottom', 'left', 'right'];
@@ -196,6 +198,12 @@ const ChannelListScreen = (props: ChannelProps) => {
                         />
                         {isTablet &&
                             <AdditionalTabletView/>
+                        }
+                        {props.showIncomingCalls && !isTablet &&
+                            <FloatingCallContainer
+                                showIncomingCalls={props.showIncomingCalls}
+                                channelsScreen={true}
+                            />
                         }
                     </Animated.View>
                 </View>
