+++ app/screens/home/channel_list/index.ts
@@ -6,6 +6,7 @@ import withObservables from '@nozbe/with-observables';
 import {of as of$} from 'rxjs';
 import {distinctUntilChanged, switchMap} from 'rxjs/operators';
 
+import {observeIncomingCalls} from '@calls/state';
 import {queryAllMyChannelsForTeam} from '@queries/servers/channel';
 import {observeCurrentTeamId, observeCurrentUserId, observeLicense} from '@queries/servers/system';
 import {queryMyTeams} from '@queries/servers/team';
@@ -23,6 +24,11 @@ const enhanced = withObservables([], ({database}: WithDatabaseArgs) => {
     );
 
     const teamsCount = queryMyTeams(database).observeCount(false);
+
+    const showIncomingCalls = observeIncomingCalls().pipe(
+        switchMap((ics) => of$(ics.incomingCalls.length > 0)),
+        distinctUntilChanged(),
+    );
 
     return {
         isCRTEnabled: observeIsCRTEnabled(database),
@@ -46,6 +52,7 @@ const enhanced = withObservables([], ({database}: WithDatabaseArgs) => {
             switchMap((u) => of$(Boolean(u))),
             distinctUntilChanged(),
         ),
+        showIncomingCalls,
     };
 });
